// src/types/schema/prisma/inputTypeSchemas/NestedFloatFilterSchema.ts
import type { Prisma } from "@prisma/client"

import { z } from "zod"

export const NestedFloatFilterSchema: z.ZodType<Prisma.NestedFloatFilter> = z.object({
  equals: z.number().optional(),
  gt: z.number().optional(),
  gte: z.number().optional(),
  in: z.number().array().optional(),
  lt: z.number().optional(),
  lte: z.number().optional(),
  not: z.union([z.number(), z.lazy(() => NestedFloatFilterSchema)]).optional(),
  notIn: z.number().array().optional(),
}).strict()

export default NestedFloatFilterSchema